services/contacts/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.32
psycopg[binary]==3.2.10
pyjwt==2.9.0
